+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [9.52.0](https://github.com/Kong/kongponents/compare/v9.51.2...v9.52.0) (2026-03-03)
+
+
+### Features
+
+* new kongponent KFilterGroup [KHCP-19257] ([#3094](https://github.com/Kong/kongponents/issues/3094)) ([2f7e81b](https://github.com/Kong/kongponents/commit/2f7e81b930246621178e98b43a4ed6c86bc9c86b)), closes [#3092](https://github.com/Kong/kongponents/issues/3092) [#3095](https://github.com/Kong/kongponents/issues/3095) [#3098](https://github.com/Kong/kongponents/issues/3098) [#3112](https://github.com/Kong/kongponents/issues/3112)
+
+## [9.51.2](https://github.com/Kong/kongponents/compare/v9.51.1...v9.51.2) (2026-03-03)
+
+
+### Bug Fixes
+
+* **kbreadcrumb:** update component styles [KHCP-19741] ([#3118](https://github.com/Kong/kongponents/issues/3118)) ([e89e3d5](https://github.com/Kong/kongponents/commit/e89e3d5101db4041b2b45d9bbf8993554ca90379))
+
+## [9.51.1](https://github.com/Kong/kongponents/compare/v9.51.0...v9.51.1) (2026-03-03)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @kong/icons to ^1.43.0 ([#3101](https://github.com/Kong/kongponents/issues/3101)) ([51230b9](https://github.com/Kong/kongponents/commit/51230b9245d3851577848098716274a6ec96c117))
+
 # [9.51.0](https://github.com/Kong/kongponents/compare/v9.50.7...v9.51.0) (2026-02-26)